pyproject.toml
[project]
name = "prepai-backend"
version = "0.1.0"
description = "PrepAI bodybuilding prep coaching backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=23.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.19.0",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.10.0",
    "openai>=1.58.0",
    "anthropic>=0.40.0",
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "testcontainers[postgres]>=4.8.0",
    "ruff>=0.9.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
markers = ["integration: requires live database"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
exclude = ["alembic/"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S", "A", "C4", "PT", "SIM"]
ignore = ["S101", "S105", "S106"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "S108", "S311"]
"alembic/**" = ["E402"]
